requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bootstrap==3.3.7.1
Werkzeug==2.2.3
opencv-python==4.7.0.72
mediapipe==0.10.21
pyttsx3==2.90
numpy==1.24.2
Pillow==9.4.0
python-dotenv==1.0.0
WTForms==3.0.1
email-validator==2.0.0
gunicorn==20.1.0
pytest==7.3.1 